chore: Rename backend package to `jobq_server`
