+++ backend/src/jobq_server/__main__.py
@@ -4,7 +4,7 @@ from contextlib import asynccontextmanager
 from fastapi import FastAPI
 from kubernetes import config
 
-from jobs_server.routers import jobs
+from jobq_server.routers import jobs
 
 
 @asynccontextmanager
